Move function_type into polymorphism
PiperOrigin-RevId: 481240161
