+++ tensorflow/core/function/polymorphism/BUILD
@@ -36,7 +36,7 @@ pytype_strict_library(
         "//tensorflow/python/eager:__subpackages__",
     ],
     deps = [
-        "//tensorflow/core/function/function_type",
+        ":function_type",
         "//tensorflow/core/function/polymorphism:type_dispatch",
         "//tensorflow/core/function/trace_type",
         "//tensorflow/python/types",
@@ -52,11 +52,37 @@ py_strict_test(
     visibility = ["//learning/brain/contrib/eager/python/examples:__pkg__"],
     deps = [
         ":function_cache",
-        "//tensorflow/core/function/function_type",
+        "//tensorflow/core/function/polymorphism:function_type",
         "//tensorflow/core/function/trace_type",
         "//tensorflow/python:array_ops",
         "//tensorflow/python/eager/polymorphic_function:function_context",
         "//tensorflow/python/platform:client_testlib",
         "//tensorflow/python/types",
+    ],
+)
+
+pytype_strict_library(
+    name = "function_type",
+    srcs = [
+        "function_type.py",
+    ],
+    srcs_version = "PY3",
+    visibility = ["//tensorflow:internal"],
+    deps = [
+        "//tensorflow/core/function/trace_type",
+        "//tensorflow/python/types",
+    ],
+)
+
+py_strict_test(
+    name = "function_type_test",
+    srcs = ["function_type_test.py"],
+    python_version = "PY3",
+    deps = [
+        ":function_type",
+        "//tensorflow/core/function/trace_type",
+        "//tensorflow/python/platform:client_testlib",
+        "//tensorflow/python/types",
+        "@absl_py//absl/testing:parameterized",
     ],
 )
