+++ tensorflow/python/eager/polymorphic_function/function_spec.py
@@ -22,7 +22,7 @@ import numpy as np
 import six
 
 from tensorflow.core.function import trace_type
-from tensorflow.core.function.function_type import function_type as function_type_lib
+from tensorflow.core.function.polymorphism import function_type as function_type_lib
 from tensorflow.python.eager.polymorphic_function import composite_tensor_utils
 from tensorflow.python.framework import composite_tensor
 from tensorflow.python.framework import constant_op
